# app/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MIDI Agent API",
    "description": "AI-powered MIDI generation from natural language prompts",
    "version": "0.1.0"
  },
  "paths": {
    "/api/midi/songs/": {
      "get": {
        "tags": ["midi"],
        "summary": "List Songs",
        "description": "List all songs for the current user.\n\nArgs:\n    db: Database session\n    user_id: User ID from Authorization header\n\nReturns:\n    List of SongResponse objects\n\nRaises:\n    HTTPException: If retrieval fails",
        "operationId": "list_songs_api_midi_songs__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SongResponse"
                  },
                  "type": "array",
                  "title": "Response List Songs Api Midi Songs  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["midi"],
        "summary": "Create Song",
        "description": "Create a new song with three default tracks (piano, bass, drums).",
        "operationId": "create_song_api_midi_songs__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSongRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/midi/songs/{song_id}": {
      "get": {
        "tags": ["midi"],
        "summary": "Get Song",
        "description": "Get a specific song with all track details and loops.\n\nArgs:\n    song_id: Song identifier (path parameter)\n    db: Database session\n    user_id: User ID from Authorization header\n\nReturns:\n    SongDetailResponse with tracks and loops\n\nRaises:\n    HTTPException: If song not found or access denied",
        "operationId": "get_song_api_midi_songs__song_id__get",
        "parameters": [
          {
            "name": "song_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Song Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/midi/loops/": {
      "post": {
        "tags": ["loops"],
        "summary": "Create Loop",
        "description": "Create a new MIDI loop.",
        "operationId": "create_loop_api_midi_loops__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLoopRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoopDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/midi/loops/{loop_id}": {
      "get": {
        "tags": ["loops"],
        "summary": "Get Loop",
        "description": "Get a specific loop with all chat messages.",
        "operationId": "get_loop_api_midi_loops__loop_id__get",
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loop Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoopDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["loops"],
        "summary": "Delete Loop",
        "description": "Delete a loop and all associated chat messages.\n\nThe chat messages are automatically deleted via cascade relationship.",
        "operationId": "delete_loop_api_midi_loops__loop_id__delete",
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loop Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["loops"],
        "summary": "Update Loop",
        "description": "Update a loop's offset, extend_measures, or track_id.\nAll fields are optional - only provided fields will be updated.",
        "operationId": "update_loop_api_midi_loops__loop_id__patch",
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchLoopRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoopResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/midi/loops/{loop_id}/chats": {
      "post": {
        "tags": ["loops"],
        "summary": "Append Chat",
        "description": "Create a new user chat message for a loop.",
        "operationId": "append_chat_api_midi_loops__loop_id__chats_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoopDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/midi/loops/{loop_id}/download": {
      "get": {
        "tags": ["loops"],
        "summary": "Download Loop Wav",
        "description": "Download a loop as a WAV file.\n\nRenders the loop to audio using the instrument samples and returns it as a downloadable WAV file.",
        "operationId": "download_loop_wav_api_midi_loops__loop_id__download_get",
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loop Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppendChatRequest": {
        "properties": {
          "loop_id": {
            "type": "string",
            "title": "Loop Id",
            "description": "ID of the loop to add the chat message to"
          },
          "msg": {
            "type": "string",
            "minLength": 1,
            "title": "Msg",
            "description": "User message content"
          },
          "measures": {
            "type": "integer",
            "maximum": 32.0,
            "minimum": 1.0,
            "title": "Measures",
            "description": "Number of measures in the loop (1-32)"
          }
        },
        "type": "object",
        "required": ["loop_id", "msg", "measures"],
        "title": "AppendChatRequest",
        "description": "Request model for appending a new chat message to a loop."
      },
...
      "CreateLoopRequest": {
        "properties": {
          "track_id": {
            "type": "string",
            "title": "Track Id",
            "description": "ID of the parent track"
          },
          "measures": {
            "type": "integer",
            "maximum": 32.0,
            "minimum": 0.0,
            "title": "Measures",
            "description": "Number of measures in the loop (1-32)"
          },
          "extend_measures": {
            "type": "integer",
            "title": "Extend Measures",
            "description": "Number of extra measures to add (or subtract) from the loop (default: 0)",
            "default": 0
          }
        },
        "type": "object",
        "required": ["track_id", "measures"],
        "title": "CreateLoopRequest",
        "description": "Request model for creating a new loop."
      },
      "CreateSongRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Song title (optional)"
          },
          "bpm": {
            "type": "integer",
            "exclusiveMaximum": 361.0,
            "exclusiveMinimum": 29.0,
            "title": "Bpm",
            "description": "Tempo in BPM (30-360)"
          },
          "key": {
            "type": "string",
            "enum": [
              "Ab",
              "A",
              "A#",
              "Bb",
              "B",
              "C",
              "C#",
              "Db",
              "D",
              "D#",
              "Eb",
              "E",
              "F",
              "F#",
              "Gb",
              "G",
              "G#",
              "Abm",
              "Am",
              "A#m",
              "Bbm",
              "Bm",
              "Cm",
              "C#m",
              "Dbm",
              "Dm",
              "D#m",
              "Ebm",
              "Em",
              "Fm",
              "F#m",
              "Gbm",
              "Gm",
              "G#m"
            ],
            "title": "Key",
            "description": "Musical key"
          },
          "time_signature": {
            "type": "string",
            "enum": ["3/4", "4/4", "5/4", "6/8", "7/8"],
            "title": "Time Signature",
            "description": "Time signature"
          }
        },
        "type": "object",
        "required": ["bpm", "key", "time_signature"],
        "title": "CreateSongRequest",
        "description": "Request payload for creating a new song."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoopDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Loop ID"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Position offset in measures from start of song"
          },
          "measures": {
            "type": "integer",
            "title": "Measures",
            "description": "Number of measures in the loop"
          },
          "extend_measures": {
            "type": "integer",
            "title": "Extend Measures",
            "description": "Number of extra measures to add (or subtract) from the loop"
          },
          "midi_events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Midi Events",
            "description": "MIDI events in the loop"
          },
          "track_id": {
            "type": "string",
            "title": "Track Id",
            "description": "ID of the parent track"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO timestamp of creation"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "ISO timestamp of last update"
          },
          "chats": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageResponse"
            },
            "type": "array",
            "title": "Chats",
            "description": "Chat messages in chronological order"
          }
        },
        "type": "object",
        "required": [
          "id",
          "offset",
          "measures",
          "extend_measures",
          "midi_events",
          "track_id",
          "created_at",
          "updated_at"
        ],
        "title": "LoopDetailResponse",
        "description": "Response model for MIDI loops, including chat message history."
      },
      "LoopResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Loop ID"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Position offset in measures from start of song"
          },
          "measures": {
            "type": "integer",
            "title": "Measures",
            "description": "Number of measures in the loop"
          },
          "extend_measures": {
            "type": "integer",
            "title": "Extend Measures",
            "description": "Number of extra measures to add (or subtract) from the loop"
          },
          "midi_events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Midi Events",
            "description": "MIDI events in the loop"
          },
          "track_id": {
            "type": "string",
            "title": "Track Id",
            "description": "ID of the parent track"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO timestamp of creation"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "ISO timestamp of last update"
          }
        },
        "type": "object",
        "required": [
          "id",
          "offset",
          "measures",
          "extend_measures",
          "midi_events",
          "track_id",
          "created_at",
          "updated_at"
        ],
        "title": "LoopResponse",
        "description": "Response model for MIDI loops, excluding chat message history."
      },
      "PatchLoopRequest": {
        "properties": {
          "offset": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Offset",
            "description": "Position offset in measures (optional)"
          },
          "extend_measures": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extend Measures",
            "description": "Number of extra measures to add (or subtract) from the loop (optional)"
          },
          "track_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Track Id",
            "description": "ID of the parent track (optional)"
          }
        },
        "type": "object",
        "title": "PatchLoopRequest",
        "description": "Request model for updating a loop."
      },
...
      "SongDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Song ID"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Song title"
          },
          "bpm": {
            "type": "integer",
            "title": "Bpm",
            "description": "Tempo in BPM"
          },
          "key": {
            "type": "string",
            "title": "Key",
            "description": "Musical key"
          },
          "time_signature": {
            "type": "string",
            "title": "Time Signature",
            "description": "Time signature"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO timestamp of creation"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "ISO timestamp of last update"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/TrackDetailResponse"
            },
            "type": "array",
            "title": "Tracks",
            "description": "Tracks in this song"
          }
        },
        "type": "object",
        "required": ["id", "title", "bpm", "key", "time_signature", "created_at", "updated_at", "tracks"],
        "title": "SongDetailResponse",
        "description": "Response model for MIDI songs including tracks and loops."
      },
      "SongResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Song ID"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Song title"
          },
          "bpm": {
            "type": "integer",
            "title": "Bpm",
            "description": "Tempo in BPM"
          },
          "key": {
            "type": "string",
            "title": "Key",
            "description": "Musical key"
          },
          "time_signature": {
            "type": "string",
            "title": "Time Signature",
            "description": "Time signature"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO timestamp of creation"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "ISO timestamp of last update"
          }
        },
        "type": "object",
        "required": ["id", "title", "bpm", "key", "time_signature", "created_at", "updated_at"],
        "title": "SongResponse",
        "description": "Response model for MIDI songs including tracks and loops."
      },
      "TrackDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Track ID"
          },
          "song_id": {
            "type": "string",
            "title": "Song Id",
            "description": "ID of the parent song"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Track title"
          },
          "midi_channel": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Midi Channel",
            "description": "MIDI channel (1-16)"
          },
          "instrument": {
            "type": "string",
            "enum": ["piano", "bass", "drum"],
            "title": "Instrument",
            "description": "Instrument type for this track"
          },
          "color": {
            "type": "string",
            "enum": ["error", "warning", "success", "info", "primary", "secondary"],
            "title": "Color",
            "description": "Color theme for this track"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO timestamp of creation"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "description": "ISO timestamp of last update"
          },
          "loops": {
            "items": {
              "$ref": "#/components/schemas/LoopResponse"
            },
            "type": "array",
            "title": "Loops",
            "description": "Loops in this track"
          }
        },
        "type": "object",
        "required": [
          "id",
          "song_id",
          "title",
          "midi_channel",
          "instrument",
          "color",
          "created_at",
          "updated_at",
          "loops"
        ],
        "title": "TrackDetailResponse",
        "description": "Response model for MIDI tracks including loops."
      },
...
    }
  }
}